report: extract independent report-generator types
