+++ lighthouse-viewer/app/src/github-api.js
@@ -11,7 +11,7 @@
 
 import idbKeyval from 'idb-keyval';
 import {FirebaseAuth} from './firebase-auth.js';
-import {getFilenamePrefix} from '../../../report/renderer/file-namer.js';
+import {getFilenamePrefix} from '../../../report/generator/file-namer.js';
 
 /**
  * Wrapper around the GitHub API for reading/writing gists.
