+++ lighthouse-viewer/types/viewer.d.ts
@@ -4,13 +4,12 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import _ReportGenerator = require('../../report/report-generator.js');
+import _ReportGenerator = require('../../report/generator/report-generator.js');
 import {DOM as _DOM} from '../../report/renderer/dom.js';
 import {ReportRenderer as _ReportRenderer} from '../../report/renderer/report-renderer.js';
 import {ReportUIFeatures as _ReportUIFeatures} from '../../report/renderer/report-ui-features.js';
 import {Logger as _Logger} from '../../report/renderer/logger.js';
 import {TextEncoding as _TextEncoding} from '../../report/renderer/text-encoding.js';
-import {getFilenamePrefix as _getFilenamePrefix} from '../../report/renderer/file-namer.js';
 import {LighthouseReportViewer as _LighthouseReportViewer} from '../app/src/lighthouse-report-viewer.js';
 import 'google.analytics';
 import {FirebaseNamespace} from '@firebase/app-types';
