+++ report/generator/flow-report-assets.js
@@ -5,4 +5,14 @@
  */
 'use strict';
 
-export {getFilenamePrefix} from '../../lighthouse-core/lib/file-namer.js';
+const fs = require('fs');
+
+/* eslint-disable max-len */
+const FLOW_REPORT_TEMPLATE = fs.readFileSync(`${__dirname}/../../flow-report/assets/standalone-flow-template.html`, 'utf8');
+const FLOW_REPORT_JAVASCRIPT = fs.readFileSync(`${__dirname}/../../dist/report/flow.js`, 'utf8');
+/* eslint-enable max-len */
+
+module.exports = {
+  FLOW_REPORT_TEMPLATE,
+  FLOW_REPORT_JAVASCRIPT,
+};
