+++ report/generator/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "commonjs",
+  "//": "Preserve commonjs in this directory. Temporary file until converted to type: module"
+}
