+++ report/generator/report-assets.js
@@ -8,9 +8,10 @@
 const fs = require('fs');
 const flowReportAssets = require('./flow-report-assets.js');
 
-const REPORT_TEMPLATE = fs.readFileSync(__dirname + '/assets/standalone-template.html', 'utf8');
-const REPORT_JAVASCRIPT = fs.readFileSync(__dirname + '/../dist/report/standalone.js', 'utf8');
-const REPORT_CSS = fs.readFileSync(__dirname + '/assets/styles.css', 'utf8');
+const REPORT_TEMPLATE = fs.readFileSync(__dirname + '/../assets/standalone-template.html',
+    'utf8');
+const REPORT_JAVASCRIPT = fs.readFileSync(__dirname + '/../../dist/report/standalone.js', 'utf8');
+const REPORT_CSS = fs.readFileSync(__dirname + '/../assets/styles.css', 'utf8');
 
 // Changes to this export interface should be reflected in build/build-dt-report-resources.js
 // and clients/devtools-report-assets.js
