+++ report/generator/tsconfig.json
@@ -1,0 +1,28 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "../../.tmp/tsbuildinfo/report/generator",
+    "emitDeclarationOnly": true,
+    "declarationMap": true,
+
+    "target": "ES2020",
+    "module": "commonjs",
+    "moduleResolution": "node",
+
+    "allowJs": true,
+    "checkJs": true,
+    "strict": true,
+    // TODO: remove the next line to be fully `strict`.
+    "useUnknownInCatchVariables": false,
+
+    // "listFiles": true,
+    // "noErrorTruncation": true,
+    "extendedDiagnostics": true,
+  },
+  "references": [
+    {"path": "../../types/lhr/"},
+  ],
+  "include": [
+    "**/*.js",
+  ],
+}
